refactor: 디렉토리별 파일 분류
