--- Frontend/k-rangers-frontend/src/css/AllPostPage.module.css
@@ -1,20 +1,0 @@
-.app {
-  display: flex;
-  justify-content: center;
-  min-height: 100vh;
-  background: #f6f7fb;
-  overflow-x: hidden;
-  width: 100%;
-}
-
-.phone {
-  width: 100%;
-  min-width: 375px;   
-  max-width: 430px;   
-  min-height: 100vh;
-  background: #ffffff;
-}
-
-.content {
- padding: 0px 20px 32px 20px;
-}
