+++ Frontend/k-rangers-frontend/src/css/Attraction/AttRecommendedList.module.css
@@ -56,4 +56,6 @@
   padding: 100px 0px;
   color: #8e8e8e;
   font-size: 16px;
+  border-radius: 25px;
+  border: 1px solid #eaeaea;
 }
