+++ Frontend/k-rangers-frontend/src/css/BottomNav.module.css
@@ -20,6 +20,7 @@
   padding-bottom: 6px;
   z-index: 1000;
   box-shadow: 0 -2px 12px rgba(0, 0, 0, 0.04);
+  border-radius: 30px 30px 0 0;
 }
 
 .navItem {
