+++ Frontend/k-rangers-frontend/src/css/PageCss/AllPostPage.module.css
@@ -1,7 +1,7 @@
 .app {
   display: flex;
   justify-content: center;
-  min-height: 100dvh;
+  min-height: 100vh;
   background: #f6f7fb;
   overflow-x: hidden;
   width: 100%;
@@ -9,13 +9,15 @@
 
 .phone {
   width: 100%;
-  min-width: 375px;  
-  max-width: 430px;
+  min-width: 375px;   
+  max-width: 430px;   
   min-height: 100vh;
   background: #ffffff;
 }
 
-.mapWrap {
-  width: 100%;
-  height: 100dvh
+.content {
+  padding: 0px 0px 32px 0px;
+  position: relative;
+  overflow-y: auto;
+  height: 100vh;
 }
