+++ Frontend/k-rangers-frontend/src/css/PageCss/MapPage.module.css
@@ -1,0 +1,59 @@
+.app {
+  display: flex;
+  justify-content: center;
+  min-height: 100dvh;
+  background: #f6f7fb;
+  overflow-x: hidden;
+  width: 100%;
+}
+
+.phone {
+  width: 100%;
+  min-width: 375px;  
+  max-width: 430px;
+  min-height: 100vh;
+  background: #ffffff;
+}
+
+.mapWrap {
+  width: 100%;
+  height: 100dvh;
+  position: relative;
+}
+
+.filterTabs {
+  position: absolute;
+  top: 20px;
+  left: 50%;
+  transform: translateX(-50%);
+  display: flex;
+  gap: 8px;
+  z-index: 10;
+  background: rgba(255, 255, 255, 0.95);
+  padding: 8px;
+  border-radius: 25px;
+  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
+  backdrop-filter: blur(10px);
+}
+
+.filterTab {
+  padding: 8px 16px;
+  border: none;
+  background: transparent;
+  border-radius: 20px;
+  font-size: 14px;
+  font-weight: 500;
+  color: #666;
+  cursor: pointer;
+  transition: all 0.2s ease;
+}
+
+.filterTab:hover {
+  background: rgba(0, 0, 0, 0.05);
+}
+
+.filterTab.active {
+  background: #007bff;
+  color: white;
+  font-weight: 600;
+}
