+++ Frontend/k-rangers-frontend/src/page/MainPage.jsx
@@ -1,9 +1,10 @@
 import React from "react";
 import Header from "../components/Header";
-import styles from "../css/MainPage.module.css";
-import HeroSection from "../components/HeroSection";
-import AccessChips from "../components/AccessChips";
-import RecommendedList from "../components/RecommendedList";
+import styles from "../css/PageCss/MainPage.module.css";
+import HeroSection from "../components/CardOption/HeroSection";
+import AccessChips from "../components/CardOption/AccessChips";
+import AttRecommendedList from "../components/Attraction/AttRecommendedList";
+import AccommodationList from "../components/Accommodation/AccomRecommendedList";
 import BottomNav from "../components/BottomNav";
 
 function MainPage() {
@@ -14,7 +15,8 @@ function MainPage() {
         <div className={styles.content}>
           <HeroSection />
           <AccessChips />
-          <RecommendedList />
+          <AttRecommendedList />
+          <AccommodationList />
         </div>
         <BottomNav />
       </div>
